# task-local xoshiro rebased