… NEWS.md
… doc/src/devdocs/subarrays.md
… doc/src/manual/performance-tips.md
… src/jltypes.c
… stdlib/LinearAlgebra/test/dense.jl
… stdlib/LinearAlgebra/test/lu.jl
… stdlib/LinearAlgebra/test/qr.jl
… stdlib/LinearAlgebra/test/uniformscaling.jl
… stdlib/Random/docs/src/index.md
… stdlib/Random/src/normal.jl
… stdlib/Random/test/runtests.jl
… stdlib/SparseArrays/src/sparsematrix.jl
… stdlib/SparseArrays/test/sparse.jl
… stdlib/Test/src/Test.jl
… stdlib/Test/test/runtests.jl
… test/error.jl
… test/threads_exec.jl