This is a cleaned-up and rebased version of #34852.

@rfourquet Please make sure I addressed your comments correctly. There is also a stack overflow in the tests due (I think) to missing overloads for the new RNG types; please advise on what should be added.

@chriselrod I haven't touched the SIMD stuff; let me know if there are specific changes I should make, or feel free to add commits here if needed.

I included a separate commit to change the default RNG to the task-local one, to see what that would look like. The main issues are
1. We don't promise the same random streams between releases, so that part is ok, but there are also APIs for e.g. copying between the global state and MersenneTwister, which could no longer be supported. Is that ok?
2. The `copy!` and `==` methods for `TaskLocal` are a bit sketchy, since `TaskLocal` is a singleton but the operations actually affect the task-local state. Do we care?
3. Single generation (`rand()`) of all types, and bulk generation of integers seem to be faster, but bulk generation of floats seems to be slower. Can this be optimized further?