@rfourquet 2021-04-21
> but there are also APIs for e.g. copying between the global state and MersenneTwister, which could no longer be supported. Is that ok?

`copy` etc. was supported, but if I'm not mistaken, neither `Random.default_rng()` nor `Random.GLOBAL_RNG` were documented or exported. Of course this was used in packages (I'm not sure to which extent), so I would say it's OK to not support that anymore (at least we are not tied by semver for this).

> The copy! and == methods for TaskLocal are a bit sketchy, since TaskLocal is a singleton but the operations actually affect the task-local state. Do we care?

I'm not sure I understand, but if you refer to `copy(::TaskLocal)` returning a `Xoshiro`, this seems fine to me, and is useful. As Kristoffer noticed very recently (probably on slack's #gripes), there is precedent of that, with `copy(ENV)` returning a `Dict`, or `copy` on views.

> Single generation (rand()) of all types, and bulk generation of integers seem to be faster, but bulk generation of floats seems to be slower. Can this be optimized further?

Should be trivially fixed, see inline comment.