@rfourquet 2021-04-21
I don't think I saw a discussion on Xoshiro256++ vs. Xoshiro256**, why was Xoshiro256++ preferred? (@chethega).