@fredrikekre 2021-04-21
For doctests there is `make -C doc html doctest=fix` to automatically fix them.