@JeffBezanson 2021-04-23
Ok there is something wrong with the simd code causing memory corruption; that's the cause of the bitarray failures.