@JeffBezanson 2021-04-23
Made some fixes to the simd code. Would be good if somebody can review.