@JeffBezanson 2021-04-23
Another unanticipated change: `MersenneTwister` saves its original seed. The test system uses that, and I can fix it by saving the seed of the global default generator, but I don't know if there are other uses we should worry about.