@rfourquet 2021-04-23
> `MersenneTwister` saves its original seed. The test system uses that, and I can fix it by saving the seed of the global default generator, but I don't know if there are other uses we should worry about.

I noticed that too, and I think that's OK without needing to save the seed of the global RNG. It's kinda breaking as this was documented, but I think that most of the benefit for the test system is kept by restoring at the end of a `@testset` the state of the global RNG as it was before the `@testset` started (this behavior is maintained in this PR). I don't think this was relied upon in packages, it's more of an interactive feature useful for debugging failing tests, and it's not too difficult to work around that (e.g. by avoiding calls to `rand()` outside of a `@testset`).