@JeffBezanson 2021-04-23
> most of the benefit for the test system is kept by restoring at the end of a `@testset` the state of the global RNG as it was before the `@testset` started (this behavior is maintained in this PR)

I agree, but then there is this evil test:
```
    a = rand()
    @testset begin
        # global RNG must re-seeded at the beginning of @testset
        @test a == rand()
```
I can just keep it working for now.