@rfourquet 2021-04-23
> I agree, but then there is this evil test:

Sure, this was testing the feature in question, so if we drop the feature we can drop the test...

> I can just keep it working for now.

You mean by saving the seed for `TaskLocalRNG` ? I though this would be out of question, as this would double the size taken for storing the RNG in each task... or did I misunderstand something?