@JeffBezanson 2021-04-23
Right. It's only tested for the case of seeding the global default RNG (`seed!(seed)`), which only requires saving a single global seed.