@rfourquet 2021-04-23
Yes this is great if this is made to work. To be explicit, the feature was meant to be useful in this context: one of your `@testset` fails, and you can reproduce the failure by doing `Random.seed!(GLOBAL_RNG.seed)` followed by including only the failed testset. An example of this in for `Base` tests, where the seed is shown in CI and you can re-run a given failed testset by copy pasting this seed. I didn't think enough about the task-local RNG setting to understand to which extent this can be replicated by storing a single seed globally.