@JeffBezanson 2021-04-23
We could even make `GLOBAL_RNG.seed` work via getproperty overloading if we want.