@oscardssmith 2021-05-01
Is that a happy (ノಠ益ಠ)ノ彡┻━┻ or a sad (ノಠ益ಠ)ノ彡┻━┻?