@JeffBezanson 2021-05-06
Ok, the qr failure can be traced to this:

```
a = [0.226633320731926 -0.2636663477154788 0.437109103001913 0.3692113076723972 0.5056807927897097; -0.6036713170198117 0.12496786793083992 0.7343834110399126 -0.8200783649587953 0.33477379709727156; 0.6110961854690776 0.7693118331381771 -0.39820685368568104 -0.03714397215871606 0.9717089309441361; -0.08575530723567074 -0.9867079454113536 0.394431400397015 -0.476587036509267 -0.21929555134741954; -0.850068401876468 0.933506636274198 -0.668391949798709 -1.0763894375665448 -0.4442798428049381; -0.2641265760084651 -0.9327717114197414 0.3975631347916196 0.4938173316477379 0.32277362111555064; 1.1000665002700676 0.9502197374692903 -0.07375477257559318 0.597051927075354 -0.24863737986566678; 0.8201589675672105 -1.099374832680121 0.3492198768383942 -0.0912919865509242 0.4237390450849754; -0.12276935472862151 -1.0982926101276103 -0.2524595052497108 -0.27657826874326047 0.432088422253561; 0.5022706893426652 -0.36463950663559885 0.7173135406029025 -0.10906174072668297 -0.2713535950049835]

factorize(a).Q
```

has different low bits on windows.