@KristofferC 2021-05-07
Can we run PkgEval on this before merging just so we have a list of packages that fail due to changes to rng?Will be useful to have when looking at PkgEval logs later when preparing the release.