@JeffBezanson 2021-05-11
I fused the minus-1.0 loop and added 11245909cdd61a28ca11632f6683d0717e2200cd. Performance looks good now:

master:
```
julia> @btime rand(100000);
  75.609 μs (2 allocations: 781.33 KiB)
julia> @btime randn(100000);
  243.082 μs (2 allocations: 781.33 KiB)
julia> @btime rand();
  9.426 ns (0 allocations: 0 bytes)
```

this PR:
```
julia> @btime rand(100000);
  49.105 μs (2 allocations: 781.33 KiB)
julia> @btime randn(100000);
  258.414 μs (2 allocations: 781.33 KiB)
julia> @btime rand();
  2.084 ns (0 allocations: 0 bytes)
```