@rfourquet 2021-05-12
Interesting, on my machine `randn` is also faster on this branch:
master:
```julia
julia> @btime rand(100000);
  154.430 μs (2 allocations: 781.33 KiB)

julia> @btime randn(100000);
  318.802 μs (2 allocations: 781.33 KiB)

julia> @btime rand();
  8.071 ns (0 allocations: 0 bytes)
```
PR:
```julia
julia> @btime rand(100000);
  68.704 μs (2 allocations: 781.33 KiB)

julia> @btime randn(100000);
  302.456 μs (2 allocations: 781.33 KiB)

julia> @btime rand();
  3.903 ns (0 allocations: 0 bytes)
```
Also `MersenneTwister` is still faster than `Xoshiro` for scalar versions, with an explicit RNG:
```julia
julia> @btime rand($(MersenneTwister()));
  2.794 ns (0 allocations: 0 bytes)

julia> @btime rand($(Xoshiro()));
  3.486 ns (0 allocations: 0 bytes)

julia> @btime rand($(MersenneTwister()), UInt);
  2.975 ns (0 allocations: 0 bytes)

julia> @btime rand($(Xoshiro()), UInt);
  3.240 ns (0 allocations: 0 bytes)
```
This is presumably due to the fact that even for scalar generation, MT uses a big cache which is filled with SIMDed array generation. So MT is still not totally superseded by `Xoshiro`, if you are willing to pay the price of such a big state.