@JeffBezanson 2021-05-19
Some interesting points from https://prng.di.unimi.it/:

- They recommend using the *upper* bits to generate floats, not the lower bits. We should probably add overloads for that for Xoshiro and modify the simd code.
- They also recommend `(x >> 11) * 0x1.0p-53` for this. That allows generating random floats with the lowest bit set. We discussed that before and concluded it wasn't worth it, but maybe worth re-evaluating now that the generator (and possibly most hardware) is faster.